refactor(cqrs):added body in its own module
